requirements.txt
accelerate==1.7.0
datasets==2.19.1
GPUtil==1.4.0
hf-xet==1.1.3
huggingface-hub==0.33.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
jupyter_client==8.6.3
jupyter_core==5.8.1
lightning-utilities==0.14.3
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
multiprocess==0.70.16
numpy==1.26.4
pandas==2.3.0
pillow==11.2.1
pyreadr==0.5.3
pytorch-lightning==2.5.1.post0
PyYAML==6.0.2
requests==2.32.4
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
sklearn-compat==0.1.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.7.1
torchmetrics==1.7.3
transformers==4.52.4
